fix: Improve token metadata update process (#11710)
* fix: Improve token metadata update process

* Add tokens metadata_updated_at column

* Allow nil metadata_updated_at

* Update test

* Order by metadata_updated_at in stream_cataloged_tokens

* token_metadata_update_on_conflict/0

* Fix metadata_updated_at usage

* Process review comment

* Process review comment
